Add drag-active property to tray control (#8604)
The drag-active property can be set to provide a highlight
when the tray control is accepting target of an ongoing drag.

We can't use the drag_highlight method because we desire to
adjust the background of the widget, rather than drawing on
top of it.  Overriding these methods would work in most cases,
but in cases where GTK+ calls drag_unhighlight itself things
could break, so we use a property instead.