Fix ContentBundle bundle_id
Most content bundles use global_name as the ID-style thing, as suggested
on the OLPC wiki. bundle_class seems undocumented and hence is not
present in any of the content bundles I have here.

Change get_bundle_id() to fall back on the global_name if no bundle_class
is set. This fixes a complete sugar crash when installing the standard
content bundles (it tried to send a None value over dbus).