configure.ac: Require gobject-introspection
configure.ac already checked if gobject-introspection is installed, but
instead of aborting with an error it continued and silently omitted crucial
parts of the build system, confusing the user with the error message

make[4]: *** No rule to make target `SugarExt-1.0.gir', needed by `all-am'.  Stop.

much later in the build. Fix this by using the appropriate autoconf macro.

Signed-off-by: Sascha Silbe <silbe@activitycentral.com>
Acked-by: Simon Schampijer <simon@laptop.org>