Implement only the test target, not check
At the moment our tests requires to be run inside a clean
dbus/xvfb session, which is not always available where the
check target is used (for example in distcheck). Also when
running distcheck in buildbot we don't really want to run
tests again, since we just run them in the check step.

We might revisit this in the future but using a non standard
target is the easiest solution for now.