Fix annotation for xdg_mime_list_parents (SL #4530)
As per SL #4530, there is a fatal crash in Sugar when retrieving mimetypes for some Journal entries when contructing the icon palette (for both the Journal list view and detail view). We isolated the problem to xdgmime.c, the call to xdg_get_mime_parents. Dnarvaez observed that the annotation for the return argument was missing and that the annotation for the return for mime_list_parents was wrong. This patch fixes these annotations.